[PATCH] FRV: Semaphore implementation race fix
This fixes a race in the FRV arch's semaphore implementation.  The same
type of fixes were applied to the rw-semaphore implementations to fix
the same races there.

The race involved the on-stack record linked into the semaphore's queue
by the down() executed by a process now sleeping on the semaphore going
away and the sleeping task going away before the process that woke it up
during up() processing had finished with those structures.

Signed-Off-By: David Howells <dhowells@redhat.com>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>