[PATCH] bonding: Update kconfig description
This patch updates the very-outdated Kconfig description for bonding.
Users reading the help text in menuconfig or xconfig would see text that
implied that bonding only supports static link aggregation, and required a
specific switch to make it work.  The new description mentions multiple
bonding modes and points the user to the bonding.txt documentation.

Signed-off-by:  Mitch Williams <mitch.a.williams@intel.com>
Signed-off-by: Jay Vosburgh <fubar@us.ibm.com>
Signed-off-by: Jeff Garzik <jgarzik@pobox.com>