[PATCH] USB: ehci requeue revisit
This gets rid of a bug found in some IRQ handling logic, after tripping
a debug assertion.  Basically, a recent patch called the wrong routine to
unlink a QH.  Net result, it wasn't allowing for the case that some other
QH was already being unlinked.  This patch uses the correct routine; the
names are confusingly similar, and the effect is often identical.

The consequence of using the wrong routine was that the driver could
lose one of the pending unlinks (probably wedging some activity) and
treat the other one as completed before it was safe to do so (which
probably wouldn't oops, but could cause other nasty corruption).

From: Brian Murphy <brian@murphy.dk>
Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>