[TCP]: Fix BIC max_cwnd calculation error.
The BIC TCP cwnd problem as identified by Yee-Ting Li and Doug Leith
is that the computation is recalc_ssthresh is incorrect and
BICTCP_1_OVER_BETA/2 should be BICTCP_1_OVER_BETA*2.

My fix is to implement the code from BIC TCP 1.1 which uses a sysctl
to set the beta.  There are a few variable name changes from the 1.1
code, and made the scaling factor a #define instead of hardcoded.

I validated this using netem and kprobes, for more details see
http://developer.osdl.org/shemminger/bic-beta-patch.pdf

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: David S. Miller <davem@davemloft.net>