[PATCH] drivers/net/wan/z85230.c interrupt handling fix
From: tom watson <lkml@tommywatson.com>

While working on a driver for z85230 based board I noticed what looks like
it could be a problem.  If the interrupt handler is handling an interrupt
on port b and an interrupt comes in for port a, it seems to me that the
port b handler would be called instead of the port a handler, and possibly
hang the board until reset.

Attached is a patch to set the irq methods back to port a before attempting
to call them.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Jeff Garzik <jgarzik@pobox.com>