[PATCH] SCSI tape: write filemark before rewind etc. when writing
The patch at the end of this message modifies the st write semantics in
the following way: write a filemark before rewind, offline, or seek if the
previous operation was write.

This semantics is specified on the man pages of some Unices and some
software (e.g., cpio) seems to assume this. The change makes sure that the
last file on the tape is properly terminated with a filemark and reading the
file does not fail at the end.

Signed-off-by: Kai Makisara <kai.makisara@kolumbus.fi>
Signed-off-by: James Bottomley <James.Bottomley@SteelEye.com>