[PATCH] ppc64: Move systemcfg out of head.S
The "systemcfg" data structure in the ppc64 kernel is something that used
to be defined to be at a hard-coded page number in the kernel image.  This
is not necessary (at least not any more) and is a possible problem with
future developements.  This patch removes that constraint, which also
simplifies various bits of assembly in head.S that were dealing with it.

This is the first step of a deeper cleanup of systemcfg definition of usage
(and ultimately removal in it's current incarnation).

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>